feat: jpa 改造
