+++ core/core-backend/src/main/java/io/dataease/exportCenter/dao/auto/entity/CoreExportDownloadTask.java
@@ -1,57 +1,29 @@
 package io.dataease.exportCenter.dao.auto.entity;
 
-import com.baomidou.mybatisplus.annotation.TableName;
-import java.io.Serializable;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import jakarta.validation.constraints.Size;
+import lombok.Getter;
+import lombok.Setter;
+import org.hibernate.annotations.Comment;
 
-/**
- * <p>
- * 下载任务列表
- * </p>
- *
- * @author fit2cloud
- * @since 2025-06-16
- */
-@TableName("core_export_download_task")
-public class CoreExportDownloadTask implements Serializable {
-
-    private static final long serialVersionUID = 1L;
-
+@Getter
+@Setter
+@Comment("下载任务列表")
+@Entity
+@Table(name = "core_export_download_task")
+public class CoreExportDownloadTask {
+    @Id
+    @Size(max = 255)
+    @Column(name = "id", nullable = false)
     private String id;
 
+    @Column(name = "create_time")
     private Long createTime;
 
+    @Column(name = "valid_time")
     private Long validTime;
 
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public Long getCreateTime() {
-        return createTime;
-    }
-
-    public void setCreateTime(Long createTime) {
-        this.createTime = createTime;
-    }
-
-    public Long getValidTime() {
-        return validTime;
-    }
-
-    public void setValidTime(Long validTime) {
-        this.validTime = validTime;
-    }
-
-    @Override
-    public String toString() {
-        return "CoreExportDownloadTask{" +
-        "id = " + id +
-        ", createTime = " + createTime +
-        ", validTime = " + validTime +
-        "}";
-    }
 }
