--- core/core-backend/src/main/java/io/dataease/exportCenter/dao/auto/mapper/CoreExportDownloadTaskMapper.java
@@ -1,18 +1,0 @@
-package io.dataease.exportCenter.dao.auto.mapper;
-
-import io.dataease.exportCenter.dao.auto.entity.CoreExportDownloadTask;
-import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-import org.apache.ibatis.annotations.Mapper;
-
-/**
- * <p>
- * 下载任务列表 Mapper 接口
- * </p>
- *
- * @author fit2cloud
- * @since 2025-06-16
- */
-@Mapper
-public interface CoreExportDownloadTaskMapper extends BaseMapper<CoreExportDownloadTask> {
-
-}
